Move common cpp app & lib tests to parent test spec
