+++ subprojects/language-native/src/integTest/groovy/org/gradle/language/cpp/CppApplicationIntegrationTest.groovy
@@ -26,7 +26,6 @@ import org.gradle.nativeplatform.fixtures.app.CppAppWithLibraryAndOptionalFeatur
 import org.gradle.nativeplatform.fixtures.app.CppAppWithOptionalFeature
 import org.gradle.nativeplatform.fixtures.app.CppCompilerDetectingTestApp
 import org.gradle.nativeplatform.fixtures.app.SourceElement
-import spock.lang.Unroll
 
 import static org.gradle.util.Matchers.containsText
 
@@ -1033,108 +1032,5 @@ class CppApplicationIntegrationTest extends AbstractCppIntegrationTest implement
 
         executable("build/exe/main/debug/app").assertExists()
         installation("build/install/main/debug").exec().out == app.expectedOutput(toolChain)
-    }
-
-    @RequiresInstalledToolChain(ToolChainRequirement.GCC_COMPATIBLE)
-    def "build succeeds when cpp source uses language features of the requested source compatibility"() {
-        given:
-        buildFile << """
-            apply plugin: 'cpp-application'
-            
-            application {
-                binaries.configureEach {
-                    sourceCompatibility = CppSourceCompatibility.Cpp11
-                }
-            }
-         """
-
-        when:
-        file("src/main/cpp/cpp11.cpp") << """
-            #include <iostream>
-            #include <ctime>
-            #include <chrono>
-            int main () {
-              using namespace std::chrono;
-              system_clock::time_point today = system_clock::now();
-              time_t tt;
-              tt = system_clock::to_time_t ( today );
-              std::cout << "today is: " << ctime(&tt);
-              return 0;
-            }
-"""
-
-        then:
-        succeeds "assemble"
-    }
-
-    @RequiresInstalledToolChain(ToolChainRequirement.GCC_COMPATIBLE)
-    def "build fails when cpp source uses language features outside the requested source compatibility"() {
-        given:
-        buildFile << """
-            apply plugin: 'cpp-application'
-            
-            application {
-                binaries.configureEach {
-                    sourceCompatibility = CppSourceCompatibility.Cpp98
-                }
-            }
-         """
-
-        when:
-        file("src/main/cpp/cpp11.cpp") << """
-            #include <iostream>
-            #include <ctime>
-            #include <chrono>
-            int main () {
-              using namespace std::chrono;
-              system_clock::time_point today = system_clock::now();
-              time_t tt;
-              tt = system_clock::to_time_t ( today );
-              std::cout << "today is: " << ctime(&tt);
-              return 0;
-            }
-"""
-
-        then:
-        fails "assemble"
-        failure.assertHasDescription("Execution failed for task ':compileDebugCpp'.")
-        failure.assertHasCause("A build operation failed.")
-        failure.assertThatCause(containsText("C++ compiler failed while compiling cpp11.cpp"))
-    }
-
-    @RequiresInstalledToolChain(ToolChainRequirement.GCC_COMPATIBLE)
-    def "recompile when source compatibility changes"() {
-        when:
-        buildFile << """
-            apply plugin: 'cpp-application'
-            
-            application {
-                binaries.configureEach {
-                    sourceCompatibility = CppSourceCompatibility.Cpp98
-                }
-            }
-         """
-
-        and:
-        file("src/main/cpp/main.cpp") << """
-            int main () {
-              return 0;
-            }
-"""
-
-        then:
-        succeeds "assemble"
-
-        when:
-        buildFile << """
-            application {
-                binaries.configureEach {
-                    sourceCompatibility = CppSourceCompatibility.Cpp11
-                }
-            }
-         """
-
-        then:
-        executedAndNotSkipped(":assemble")
     }
 }
